Rename auth service to fit convention and change background theme
