+++ lib/frontend/pages/login_page.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:project_fomo/frontend/pages/root_page.dart';
 import 'package:provider/provider.dart';
 import 'package:project_fomo/backend/state_models/login_model.dart';
-import 'package:project_fomo/backend/services/AuthService.dart';
+import 'package:project_fomo/backend/services/auth_service.dart';
 
 //TODO: Recreate login page UI
 class LoginPage extends StatelessWidget {
