+++ lib/frontend/pages/register_page.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'package:project_fomo/backend/state_models/register_model.dart';
-import 'package:project_fomo/backend/services/AuthService.dart';
+import 'package:project_fomo/backend/services/auth_service.dart';
 
 //TODO: Recreate register page UI
 class RegisterPage extends StatelessWidget {
